Only minor changes:
* Tamara's external email address in gemspec
* Add mode line in rails_template.rb.tt
* Fuzz the passwords in group_vars/staging

A concern is what to do about the staging vault. We should probably
delete it from the public repo with a note about creating it. What do
you think? It doesn't currently contain anything interesting anymore,
I removed a couple of passwords that were in it. It's still
password-protected as it is so I don't think they'll leak.